Mudando pedzap para deeliv ai e iorion19 para degusta ai
