+++ MarketPlace/Deeliv/Domain/cliente.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Deeliv.Domain
+{
+    public class cliente
+    {               
+        public string cli_nome { get; set; }
+        public string cli_telefone { get; set; }
+        public string cli_datacadastral { get; set; }        
+    }
+}
