+++ MarketPlace/Deeliv/Domain/erro_result.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Deeliv.Domain
+{
+    public class erro_result
+    {
+        public bool erro { get; set; }
+        public string descricao { get; set; }
+    }
+}
