+++ MarketPlace/Deeliv/Enum/FormaPagamentoTipo.cs
@@ -1,0 +1,54 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Deeliv.Enum
+{
+    public class FormaPagamentoTipo
+    {
+        public const string INDEFINIDO = "indefinido";        
+        public const string DINHEIRO = "dinheiro";
+        public const string CHEQUE = "cheque";
+        public const string BOLETO = "boleto";
+        public const string VOUCHER = "voucher";
+        public const string CARTAO_DEBITO = "cartao_debito";
+        public const string CARTAO_VISA = "cartao_visa";
+        public const string CARTAO_MASTERCARD = "cartao_mastercard";
+        public const string CARTAO_AMEX = "cartao_amex";
+        public const string CARTAO_ELO = "cartao_elo";
+        public const string CARTAO_DINERS = "cartao_diners";
+        public const string CARTAO_DISCOVER = "cartao_discover";
+        public const string CARTAO_JCB = "cartao_jcb";
+        public const string CARTAO_AURA = "cartao_aura";
+        public const string CARTAO_HIPERCARD = "cartao_hipercard";
+        public const string CARTAO_BANRICOMPRAS = "cartao_banricompras";
+        public const string CARTAO_GOODCARD = "cartao_goodcard";
+        public const string CARTAO_CREDISHOP = "cartao_credishop";
+        public const string CARTAO_AVANCARD = "cartao_avancard";
+        public const string CARTAO_CABAL = "cartao_cabal";
+        public const string PICPAY_OFFLINE = "picpay_offline";
+        public const string VALE_ALELOA_ALIMENTACAO = "vale_aleloalimentacao";
+        public const string VALE_ALELOA_REFEICAO = "vale_alelorefeicao";
+        public const string VALE_BENA_ALIMENTACAO = "vale_benalimentacao";
+        public const string VALE_BEN_REFEICAO = "vale_benrefeicao";
+        public const string VALE_COOPERCARD = "vale_coopercard";
+        public const string VALE_GREENCARD = "vale_greencard";
+        public const string VALE_REFEISUL = "vale_refeisul";
+        public const string VALE_SODEXO_ALIMENTACAO = "vale_sodexoalimentacao";
+        public const string VALE_SODEXO_REFEICAO = "vale_sodexorefeicao";
+        public const string VALE_TICKET_RESTAURANTE = "vale_ticketrestaurante";
+        public const string VALE_USECREDA_ALIMENTACAO = "vale_usecredalimentacao";
+        public const string VALE_USECREDA_REFEICAO = "vale_usecredrefeicao";
+        public const string VALE_VALECARD = "vale_valecard";
+        public const string VALE_VEROCARD_ALIMENTACAO = "vale_verocardalimentacao";
+        public const string VALE_VEROCARD_REFEICAO = "vale_verocardrefeicao";
+        public const string VALE_VR_ALIMENTACAO = "vale_vralimentacao";
+        public const string VALE_VR_REFEICAO = "vale_vrrefeicao";
+        public const string PONTOS_ACUMULADOS = "pontos_acumulados";
+        public const string MERCADO_PAGO_CARTAO = "mercadopago_cartao";
+        public const string PAGSEGURO = "pagseguro";
+        public const string PICPAY_ONLINE = "picpay_online";
+    }
+}
