+++ MarketPlace/Deeliv/Enum/PedidoOrigem.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Deeliv.Enum
+{
+    public class PedidoOrigem
+    {
+        public const string PAI = "PAI";
+        public const string WEB = "WEB";
+        public const string APP = "APP";
+        public const string WHA = "WHA";
+        public const string TEL = "TEL";
+        public const string MES = "MES";        
+    }
+}
