+++ MarketPlace/Deeliv/Enum/PerguntaFormato.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Deeliv.Enum
+{
+    public class PerguntaFormato
+    {
+        public const string UNICA = "UNI";
+        public const string MULTIPLA = "MUL";
+        public const string DIVISORIA = "DIV";
+        public const string LIVRE = "LIV";        
+    }
+}
