+++ MarketPlace/Deeliv/Properties/AssemblyInfo.cs
