+++ MarketPlace/DegustaAi/Domain/endereco_de_entrega.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DegustaAi.Domain
+{
+    public class endereco_de_entrega
+    {
+        public string logradouro { get; set; }
+        public string numero { get; set; }
+        public string complemento { get; set; }
+        public string referencia { get; set; }
+        public string bairro { get; set; }
+        public string cep { get; set; }
+        public string cidade { get; set; }
+        public string uf { get; set; }
+    }
+}
