+++ MarketPlace/DegustaAi/Enum/FormaPagamento.cs
@@ -1,0 +1,48 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DegustaAi.Enum
+{
+    public enum FormaPagamento
+    {
+        ValeSodexo = 1,
+        ValeVRSmart = 2,
+        VisaRefeicao = 3,
+        CreditoMasterCard = 4,
+        CreditoElo = 5,
+        Dinheiro = 6,
+        Cheque = 7,
+        ValeVerocard = 8,
+        NutricardRefeicaoAlimentacao = 9,
+        ValeAleloRefeicao = 10,
+        ValeRefeisul = 11,
+        ValeGreenCardPapel = 12,
+        ValeFreenCardCaartao = 13,
+        ValeTicketRestaurante = 14,
+        ValeCard = 15,
+        BenVisaRefeicao = 16,
+        ValeCooperCard = 17,
+        CreditoDiners = 18,
+        CreditoGoodCard = 19,
+        DebitoElo = 20,
+        CreditoVisa = 21,
+        CreditoHipercard = 22,
+        CreditoBanricompras = 23,
+        DebitoVisa = 24,
+        CreditoAmericanExpress = 25,
+        CreditoNugo = 26,
+        CreditoVerdecard = 27,
+        DebitoMastercard = 28,
+        PicPay = 29,
+        CartaoCreditoOnline = 30,
+        CartaoDebitoOnline = 31,
+        DebitoBanricompras = 32,
+        Ame = 33,
+        Pix = 35,
+        TransferenciaBancaria = 36,
+        Papayas = 37,
+        WhatsAppPay = 38,
+        PixOnline = 51
+    }
+}
