+++ MarketPlace/DegustaAi/Enum/Status.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DegustaAi.Enum
+{
+    public enum Status
+    {
+        PedidoCancelado = 0,
+        PedidoPendente = 1,
+        PedidoEmProducao = 2,
+        PedidoEmTransitoOuProntoParaRetirada = 3,
+        PeditoEntregueOuRetirado = 4,
+        PedidoAceito_EmCasoDeAgendamento = 5
+    }
+}
