+++ MarketPlace/Example/Example.csproj
@@ -141,6 +141,14 @@
       <Project>{90c857a2-55f0-4314-b9fc-a3c9b40add13}</Project>
       <Name>CRMBonus</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Deeliv\Deeliv.csproj">
+      <Project>{9a45d340-8699-4fc2-9264-d5514b6d511d}</Project>
+      <Name>Deeliv</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\DegustaAi\DegustaAi.csproj">
+      <Project>{602a6b91-7846-4468-a51e-505b54086a15}</Project>
+      <Name>DegustaAi</Name>
+    </ProjectReference>
     <ProjectReference Include="..\DeliveryApp\DeliveryApp.csproj">
       <Project>{b95f22fb-e688-4a01-946a-5daf2d5c1457}</Project>
       <Name>DeliveryApp</Name>
@@ -185,10 +193,6 @@
       <Project>{72ee52a9-3a7a-43a6-b8e6-b8a10b05f05e}</Project>
       <Name>Ifood</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Iorion19\Iorion19.csproj">
-      <Project>{602a6b91-7846-4468-a51e-505b54086a15}</Project>
-      <Name>Iorion19</Name>
-    </ProjectReference>
     <ProjectReference Include="..\JotaJa\JotaJa.csproj">
       <Project>{e50b2095-11b2-4926-bc46-3b0b6bb99899}</Project>
       <Name>JotaJa</Name>
@@ -212,10 +216,6 @@
     <ProjectReference Include="..\OnPedido\OnPedido.csproj">
       <Project>{f423ea0e-b5c2-4152-ac20-b2de04b4b1aa}</Project>
       <Name>OnPedido</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\PedZap\PedZap.csproj">
-      <Project>{9a45d340-8699-4fc2-9264-d5514b6d511d}</Project>
-      <Name>PedZap</Name>
     </ProjectReference>
     <ProjectReference Include="..\PixCommerce\PixCommerce.csproj">
       <Project>{dacbf174-3fbe-4a18-b22a-15ff999c7807}</Project>
