--- MarketPlace/Iorion19/Domain/response.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Iorion19.Domain
-{
-    public class response
-    {
-        public response()
-        {
-            pedidos = new List<pedido>();
-        }
-
-        public string status { get; set; }
-        public string code { get; set; }
-        public int pedidos_count { get; set; }
-        public List<pedido> pedidos { get; set; }
-        public int status_pedido {  get; set; }
-        public string message { get; set; }
-    }
-}
