--- MarketPlace/Iorion19/Utils/Constants.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Iorion19.Utils
-{
-    public static class Constants
-    {
-        public const string URL_GET_ORDERS = "/api/delivery/getorders";
-        public const string URL_CHANGE_STATUS = "/api/delivery/changestatus";
-        public const string URL_CANCELA_PEDIDO = "/api/delivery/cancelaPedido";
-        public const string URL_SEND_TO_MOTOBOY_SERVICE = "/api/delivery/sendToMotoboyService";
-    }
-}
