--- MarketPlace/PedZap/Domain/cliente.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PedZap.Domain
-{
-    public class cliente
-    {               
-        public string cli_nome { get; set; }
-        public string cli_telefone { get; set; }
-        public string cli_datacadastral { get; set; }        
-    }
-}
