--- MarketPlace/PedZap/Domain/erro_result.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PedZap.Domain
-{
-    public class erro_result
-    {
-        public bool erro { get; set; }
-        public string descricao { get; set; }
-    }
-}
