--- MarketPlace/PedZap/Enum/PedidoOrigem.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PedZap.Enum
-{
-    public class PedidoOrigem
-    {
-        public const string PAI = "PAI";
-        public const string WEB = "WEB";
-        public const string APP = "APP";
-        public const string WHA = "WHA";
-        public const string TEL = "TEL";
-        public const string MES = "MES";        
-    }
-}
