--- MarketPlace/PedZap/Enum/PedidoStatus.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PedZap.Enum
-{
-    public class PedidoStatus
-    {
-        public const string ABERTO = "ABE";
-        public const string AJUDA = "AJU";
-        public const string PENDENTE = "PEN";
-        public const string ACEITO = "ACE";
-        public const string PREPARADO = "PRE";
-        public const string NO_LOCAL = "LOC";
-        public const string ENTREGUE = "ENT";
-        public const string RETIRADO = "RET";
-        public const string AVALIADO = "AVA";
-        public const string REJEITADO = "REJ";
-        public const string REGIAO_NAO_ATENDIDA = "REG";
-        public const string DESISTENCIA = "DES";        
-    }
-}
