--- MarketPlace/PedZap/Enum/PedidoTipo.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PedZap.Enum
-{
-    public class PedidoTipo
-    {
-        public const string BALCAO = "BAL";
-        public const string MESA = "MES";
-        public const string RETIRADA = "RET";
-        public const string ENTREGA = "ENT";       
-    }
-}
