--- MarketPlace/PedZap/Utils/Constants.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PedZap.Utils
-{
-    public static class Constants
-    {
-        //public const string URL_BASE = "https://www.pedzap.com.br/apiv1/";
-        public const string URL_BASE = "https://www.deeliv.app/apiv1/";
-
-        public const string URL_PEDIDOS = "pedidos";
-        public const string URL_PEDIDO = "pedido";
-        public const string URL_PEDIDOS_ITENS = "pedidos_itens";
-        public const string URL_PEDIDOS_ITENS_PERGUNTAS = "pedidos_itens_perguntas";
-        public const string URL_PEDIDOS_ITENS_RESPOSTAS = "pedidos_itens_respostas";
-        public const string URL_PEDIDOS_STATUS = "pedidos_status";
-        public const string URL_CLIENTE = "cliente";
-    }
-}
-
